+++ rearq/__init__.py
@@ -110,6 +110,7 @@ class ReArq:
         name: Optional[str] = None,
         job_retry: Optional[int] = None,
         job_retry_after: Optional[int] = None,
+        job_timeout: Optional[int] = None,
         expire: Optional[Union[float, datetime.datetime]] = None,
         run_at_start: Optional[Union[bool, Tuple, Dict]] = False,
         run_with_lock: bool = False,
@@ -128,6 +129,7 @@ class ReArq:
             rearq=self,
             job_retry=job_retry or self.job_retry,
             job_retry_after=job_retry_after or self.job_retry_after,
+            job_timeout=job_timeout or self.job_timeout,
             expire=expire or self.expire,
             bind=bind,
             run_with_lock=run_with_lock,
@@ -151,6 +153,7 @@ class ReArq:
         name: Optional[str] = None,
         job_retry: Optional[int] = None,
         job_retry_after: Optional[int] = None,
+        job_timeout: Optional[int] = None,
         expire: Optional[Union[float, datetime.datetime]] = None,
         run_at_start: Optional[Union[bool, Tuple, Dict]] = False,
         run_with_lock: bool = False,
@@ -163,6 +166,7 @@ class ReArq:
         :param name: Task name, default is function name.
         :param job_retry: Override default job retry.
         :param job_retry_after: Override default job retry after.
+        :params: job_timeout: Override default job timeout.
         :param expire: Override default expire.
         :param run_at_start: Whether run at startup or not, only work timer worker
         :param run_with_lock: Run task with redis lock, only one task can run at the same time if enabled.
@@ -178,6 +182,7 @@ class ReArq:
                 name,
                 job_retry,
                 job_retry_after,
+                job_timeout,
                 expire,
                 run_at_start,
                 run_with_lock,